+++ example/lib/login_form_example.dart
@@ -33,9 +33,7 @@ class _LoginFormExampleState extends State<LoginFormExample> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(
-        title: const Text('Login Form Example'),
-      ),
+      appBar: AppBar(title: const Text('Login Form Example')),
       body: Padding(
         padding: const EdgeInsets.all(16.0),
         child: Form(